Starting optimizing bv.py, topaz is getting unstable, moving to local machines
