+++ main.py
@@ -4,8 +4,8 @@ from io import xsf_info, el_info
 from io import qe_out_info, make_qe_in
 from io import init_log, upd_log
 from io import init_axsf, upd_axsf
-from bv import bv
 from mc import mc
+from bv import bv
 import os
 import numpy as np
 
@@ -13,31 +13,29 @@ xsf_filename = sys.argv[1] # read xsf filename from command line
 el_filename = sys.argv[2] # read element list filename
 
 # set simulation parameters
-niter = 50000
+niter = 100
 max_disp = 0.05 # angstroms
 T_move = 1 # kelvin
-T_exc = T_move # temperature for exchange steps, use with caution
 ry_ev = 13.605693009
 bohr_ang = 0.52917721067
 buf_len = 2.0 # length above surface within which atoms can be added
 mu_list = [0, 0, 0] # sr, ti, o
-act_p = np.array([1,1,1,0,0]) # probablity of taking different actions, [0]: move, [1]: swap, [2]: add, [3]: remove
+act_p = np.array([1,0,0,1,1]) # probablity of taking different actions, [0]: move, [1]: swap, [2]: jump, [3]: add, [4]: remove
 
 # get element info
-el = el_info(el_filename) # instantiates el_info object
-el.pop_attr(T_exc) # populate attributes in el_info object
+el = el_info() # instantiates el_info object
+el.pop_attr(el_filename,T_move)
 
 # get info from xsf file
-xsf = xsf_info(xsf_filename) # instantiate xsf_info objects
-xsf_old = xsf_info(xsf_filename)
-xsf_new = xsf_info(xsf_filename)
-xsf.pop_attr(buf_len, el) # populate attributes in xsf_info object
+xsf = xsf_info() # instantiate xsf_info objects
+xsf.pop_attr(xsf_filename, el, buf_len) # populate attributes in xsf_info object
 
 # create bv object
 bvo = bv()
 
 # instantiate mc object
-mc_test = mc(T_move, T_exc, max_disp, xsf)
+mc_run = mc()
+mc_run.init(T_move, max_disp, xsf)
 
 # run mc simulation
 os.system('mkdir -p temp') # make temp directory for qe calculations
@@ -45,51 +43,43 @@ os.chdir('temp') # enter temp
 log_file = init_log('log.dat') # initialize log file
 axsf_file = init_axsf('coord.axsf', niter, xsf) # initialize axsf file
 for i in range(niter) :
-    # makes a copy of xsf attributes called xsf_old
-    xsf_old.copy(xsf)
+	# attempt uvt action and store xsf attributes in xsf_new
+	xsf = mc_run.uvt_new_structure(xsf, el, act_p, bvo) 
 
-    # attempt uvt action and store xsf attributes in xsf_new
-    xsf_new.at_coord, \
-        xsf_new.ind_rem_at, \
-        xsf_new.el_list, \
-        xsf_new.num_each_el, \
-        xsf_new.num_at = mc_test.uvt_new_structure(xsf, el, act_p, bvo) 
-    
-    # copy xsf_new attributes to xsf, try to figure out a way around this
-    xsf.copy(xsf_new)
+	# make input file
+	make_qe_in('qe.in', xsf, el)
+	
+#	# calculate and get total energy
+#	if xsf.num_at <= 2 :
+#		os.system('mpiexec.hydra -np 4 ../bin/pw.x -i qe.in > qe.out') # execute qe
+#	elif xsf.num_at <= 6 :
+#		os.system('mpiexec.hydra -np 9 ../bin/pw.x -i qe.in > qe.out')
+#	else :
+#		os.system('mpiexec.hydra -np 36 ../bin/pw.x -i qe.in > qe.out')
+#	qe_out = qe_out_info('qe.out')
+#	new_en = qe_out.get_final_en() * ry_ev # convert final energy from ry to ev
+#	forces = qe_out.get_forces() * ry_ev / bohr_ang # convert forces from ry/bohr to ev/ang
+	new_en = - i
+	forces = np.zeros((xsf.at_num,3))
 
-    # make input file
-    make_qe_in('qe.in', xsf)
-    
-    # calculate and get total energy
-    if xsf.num_at <= 2 :
-        os.system('mpiexec.hydra -np 4 ../bin/pw.x -i qe.in > qe.out') # execute qe
-    elif xsf.num_at <= 6 :
-        os.system('mpiexec.hydra -np 9 ../bin/pw.x -i qe.in > qe.out')
-    else :
-        os.system('mpiexec.hydra -np 36 ../bin/pw.x -i qe.in > qe.out')
-    qe_out = qe_out_info('qe.out')
-    new_en = qe_out.get_final_en() * ry_ev # convert final energy from ry to ev
-    forces = qe_out.get_forces() * ry_ev / bohr_ang # convert forces from ry/bohr to ev/ang
+	# update T
+	mc_run.update_T_const(i, 3000)
 
-    # update T
-    mc_test.update_T_const(T_move, i, 3000)
+	# decide whether or not to accept uvt action, 
+	accept = mc_run.uvt_mc(new_en, el, mu_list)
 
-    # decide whether or not to accept uvt action, 
-    accept = mc_test.uvt_mc(new_en, xsf, el, mu_list)
+	# calculate free energy 
+	free_en, _ = mc_run.get_free_g_p(new_en, el, mu_list)
 
-    # calculate free energy 
-    free_en, _ = mc_test.get_free_g_p(new_en, xsf, el, mu_list)
+	# if step not accepted, copy attributes from xsf_old to xsf
+	if accept == 0 :
+		xsf = mc_run.old_xsf.copy()
 
-    # if step not accepted, copy attributes from xsf_old to xsf
-    if accept == 0 :
-        xsf.copy(xsf_old)
+	# write energies, number of accepted steps, and acceptance rate to log file
+	upd_log(log_file, i, free_en, mc_run)
 
-    # write energies, number of accepted steps, and acceptance rate to log file
-    upd_log(log_file, i, free_en, mc_test)
-
-    # write atomic coordinates to axsf file
-    upd_axsf(axsf_file, i, xsf, forces)
+	# write atomic coordinates to axsf file
+	upd_axsf(axsf_file, i, xsf, el, forces)
 
 log_file.close()
 axsf_file.close()
