--- new_main.py
@@ -1,86 +1,0 @@
-import sys
-import copy
-from new_io import xsf_info, el_info
-from new_io import qe_out_info, make_qe_in
-from new_io import init_log, upd_log
-from new_io import init_axsf, upd_axsf
-from new_mc import mc
-from new_bv import bv
-import os
-import numpy as np
-
-xsf_filename = sys.argv[1] # read xsf filename from command line
-el_filename = sys.argv[2] # read element list filename
-
-# set simulation parameters
-niter = 100
-max_disp = 0.05 # angstroms
-T_move = 1 # kelvin
-ry_ev = 13.605693009
-bohr_ang = 0.52917721067
-buf_len = 2.0 # length above surface within which atoms can be added
-mu_list = [0, 0, 0] # sr, ti, o
-act_p = np.array([1,1,1,0,0]) # probablity of taking different actions, [0]: move, [1]: swap, [2]: add, [3]: remove
-
-# get element info
-el = el_info() # instantiates el_info object
-el.pop_attr(el_filename,T_move)
-
-# get info from xsf file
-xsf = xsf_info() # instantiate xsf_info objects
-xsf.pop_attr(xsf_filename, el, buf_len) # populate attributes in xsf_info object
-
-# create bv object
-bvo = bv()
-
-# instantiate mc object
-mc_run = mc()
-mc_run.init(T_move, max_disp, xsf)
-
-# run mc simulation
-os.system('mkdir -p temp') # make temp directory for qe calculations
-os.chdir('temp') # enter temp
-log_file = init_log('log.dat') # initialize log file
-axsf_file = init_axsf('coord.axsf', niter, xsf) # initialize axsf file
-for i in range(niter) :
-	# attempt uvt action and store xsf attributes in xsf_new
-	xsf = mc_run.uvt_new_structure(xsf, el, act_p, bvo) 
-
-	# make input file
-	make_qe_in('qe.in', xsf, el)
-	
-#	# calculate and get total energy
-#	if xsf.num_at <= 2 :
-#		os.system('mpiexec.hydra -np 4 ../bin/pw.x -i qe.in > qe.out') # execute qe
-#	elif xsf.num_at <= 6 :
-#		os.system('mpiexec.hydra -np 9 ../bin/pw.x -i qe.in > qe.out')
-#	else :
-#		os.system('mpiexec.hydra -np 36 ../bin/pw.x -i qe.in > qe.out')
-#	qe_out = qe_out_info('qe.out')
-#	new_en = qe_out.get_final_en() * ry_ev # convert final energy from ry to ev
-#	forces = qe_out.get_forces() * ry_ev / bohr_ang # convert forces from ry/bohr to ev/ang
-	new_en = - i
-	forces = np.zeros((xsf.at_num,3))
-
-	# update T
-	mc_run.update_T_const(i, 3000)
-
-	# decide whether or not to accept uvt action, 
-	accept = mc_run.uvt_mc(new_en, el, mu_list)
-
-	# calculate free energy 
-	free_en, _ = mc_run.get_free_g_p(new_en, el, mu_list)
-
-	# if step not accepted, copy attributes from xsf_old to xsf
-	if accept == 0 :
-		xsf = mc_run.old_xsf.copy()
-
-	# write energies, number of accepted steps, and acceptance rate to log file
-	upd_log(log_file, i, free_en, mc_run)
-
-	# write atomic coordinates to axsf file
-	upd_axsf(axsf_file, i, xsf, el, forces)
-
-log_file.close()
-axsf_file.close()
-os.chdir('../')
