+++ structure.xsf
@@ -16,7 +16,7 @@ O   0.000 -1.975 -1.975 1
 Sr  5.925  1.975  1.975 1
 O   5.925  0.000  1.975 1
 O   3.95  1.975  1.975 1
-Ti  3.95  0.000  0.000 1
+Ti  3.95  0.000  0.000 0
 O   1.975  1.975  3.95 1
 Ti  1.975 -1.975 -1.975 1
 O   1.975  0.000 -1.975 1
